Add SearchComponent
