+++ packages/outside/src/PageContainer/Admin/NewUserListPage/index.tsx
@@ -33,7 +33,7 @@ const NewUserListPage = () => {
       </MainStyle.SlideBg>
       <MainStyle.MainWrapper>
         <MainStyle.MainContainer>
-          <NewUserList />          
+          <NewUserList />
         </MainStyle.MainContainer>
       </MainStyle.MainWrapper>
     </MainStyle.PageWrapper>
