+++ shared/common/src/components/InquiryContainer/index.tsx
@@ -20,13 +20,12 @@ const InquiryContainer = ({ isAdmin }: { isAdmin: boolean }) => {
 
   return (
     <>
-      <InquirySearch
-        isAdmin={isAdmin}
+      {isAdmin && <InquirySearch
         refetch={() => refetch()}
         keyword={keyword}
         setKeyword={setKeyword}
         setAnswerStatus={setAnswerStatus}
-      />
+      />}
       <InquiryList inquiryList={inquiryList} />
     </>
   )
