+++ shared/common/src/components/index.ts
@@ -16,6 +16,7 @@ export { default as PaginationPages } from './PaginationPages'
 export { default as PostItem } from './PostItem'
 export { default as PrivateRouter } from './PrivateRouter'
 export { default as SchoolIntro } from './SchoolIntro'
+export { default as SearchComponent } from './SearchComponent'
 export { default as Sequence } from './Sequence'
 export * from './Sliders'
 export { default as UserItem } from './UserItem'
