+++ shared/common/src/styles/pageLayout.ts
@@ -48,6 +48,7 @@ export const SlideButton = styled.div`
   padding: 0 0.75rem;
   border-radius: 0.5rem;
   backdrop-filter: blur(0.25rem);
+  
   svg {
     fill: ${({ theme }) => theme.color.white};
   }
